Create CrytoUtils, move sha256hash160 from Utils
